refactor: break shadow replay analysis cycle
