+++ docs/dependency_graph.mmd
@@ -1,5 +1,5 @@
 flowchart LR
-  src_application["src.application"] -->|165| domain_domain["domain.domain"]
+  src_application["src.application"] -->|170| domain_domain["domain.domain"]
   src_interfaces["src.interfaces"] -->|103| src_application["src.application"]
   src_application["src.application"] -->|73| src_infrastructure["src.infrastructure"]
   src_application_ledger["src.application.ledger"] -->|28| domain_domain["domain.domain"]
