+++ docs/reviews/code-review-20260723-024851.md
@@ -1,0 +1,67 @@
+# Code Review
+
+## Scope
+
+- Mode: current changes
+- Branch: `codex/close-advice-strategy-optimization`
+- Base: `HEAD@6a81789d`
+- Slice: post-merge integration of S1-S5 with `origin/main`
+- Included: Shadow Replay analysis ownership, generated dependency graph,
+  candidate-analysis import compatibility, legacy-reference inventory, and
+  post-merge regression evidence
+- Excluded: evidence collection, CEO policy selection, S6 production
+  promotion, runtime config, and notification behavior
+- Review result: no unresolved material findings after the fixes below
+
+## Findings
+
+### 1 — fixed — high — optional close analysis introduced a production import cycle
+
+- **Path**: `shadow_replay.analysis` -> `shadow_replay.readiness` ->
+  `shadow_replay.capture` -> `shadow_replay.analysis`.
+- **Failure**: attaching S4 close-policy analysis to the existing candidate
+  analysis module made that module both a reusable candidate primitive and a
+  higher-level close-analysis orchestrator. The generated dependency guard
+  detected one production module cycle after the current `origin/main` graph
+  was regenerated.
+- **Fix**: moved the pure candidate analysis implementation to
+  `shadow_replay.candidate_analysis`; internal candidate consumers now depend
+  on that lower-level module. `shadow_replay.analysis` remains the compatible
+  public facade and owns only dataset-level composition, including the
+  optional close facet.
+- **Coverage**: production dependency graph reports zero module cycles and no
+  boundary violations; candidate-only and optional close-facet analysis tests
+  pass through the public facade.
+
+### 2 — fixed — medium — moving candidate analysis invalidated the explicit legacy-reference owner path
+
+- **Path**: candidate analysis move -> performance legacy-reference scanner ->
+  explicit allowlist.
+- **Failure**: the scanner correctly reported the new file as unowned and the
+  old file path as stale even though the same candidate-domain calculation had
+  only moved ownership.
+- **Fix**: replaced the old allowlist path with the new owning module while
+  preserving the existing `candidate_strategy_domain` category. No legacy
+  term or additional file was allowlisted.
+- **Coverage**: exact allowlist inventory and stale-entry tests pass.
+
+## Verification
+
+- Close Advice, Shadow Replay, facade compatibility, and dependency graph:
+  `282 passed`, `10 skipped`.
+- Full repository suite after the fixes: `3107 passed`, `10 skipped`, six
+  pre-existing deprecation warnings.
+- Generated dependency graph: 483 production modules, zero module cycles, no
+  forbidden boundary imports.
+- Ruff on all changed Python modules: passed.
+- `git diff --check`: passed.
+
+## Residual Risk
+
+- `shadow_replay.analysis` intentionally re-exports the existing public
+  candidate-analysis names while routing internal consumers directly to the
+  new owner. Known public imports (`analyze_rows`, `_evidence_level`, and
+  `analyze_shadow_replay_dataset`) are covered by the existing suite.
+- This integration review does not change the S5 stop decision: there is still
+  no admissible prospective close-decision cohort, so P1/P2/P3 remain
+  shadow-only and production authority remains P0.
