+++ src/application/strategy_lab/experiment.py
@@ -4,7 +4,7 @@ from collections import Counter, defaultdict
 from pathlib import Path
 from typing import Any
 
-from src.application.shadow_replay.analysis import analyze_rows
+from src.application.shadow_replay.candidate_analysis import analyze_rows
 from src.application.shadow_replay.candidate_impact import (
     _enrich_iv_rv_history_percentiles,
     _evaluate_candidate,
